+++ SORA_offloader/flow.py
@@ -3,32 +3,32 @@
 
 import json
 
-def toRouterPort(switchId, port, routerPort, pairVlan):
+def toRouterPort(switchId, port, routerPort, pairVlan,tableId):
 	vlan = pairVlan + 4096
 	rule = {}
 	rule['dpid'] = switchId[0]
-	rule['table_id'] = 100
+	rule['table_id'] = tableId
 	rule['match'] = { 'in_port': port }
 	rule['actions'] = []
 	rule['actions'].append( { 'type':'SET_FIELD', 'field':'vlan_vid', 'value':vlan } )
 	rule['actions'].append( { 'type':'OUTPUT', 'port':routerPort } )
 	return rule
 
-def fromRouterPort(switchId, port, routerPort, pairVlan):
+def fromRouterPort(switchId, port, routerPort, pairVlan,tableId):
 	vlan = pairVlan + 4096
 	rule = {}
 	rule['dpid'] = switchId[0]
-	rule['table_id'] = 100
+	rule['table_id'] = tableId
 	rule['match'] = { 'in_port': routerPort, 'dl_vlan':vlan }
 	rule['actions'] = []
 	rule['actions'].append( {'type':'POP_VLAN'} )
 	rule['actions'].append( { 'type':'OUTPUT', 'port':port } )
 	return rule
 
-def offload(switchId,ipv4Prefix,nextHopMac,nextHopPort,timeout):
+def offload(switchId,ipv4Prefix,nextHopMac,nextHopPort,timeout,tableId):
 	rule = {}
 	rule['dpid'] = switchId[0]
-	rule['table_id'] = 100 
+	rule['table_id'] = tableId 
 	rule['priority'] = 500 
 	rule['hard_timeout'] = timeout  
 	rule['match'] = { 'ipv4_dst':ipv4Prefix, 'eth_type':2048 }
